% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_train.R
\name{majority_class}
\alias{majority_class}
\title{Majority Class}
\usage{
majority_class(target)
}
\arguments{
\item{target}{A vector of class labels.}
}
\value{
The majority class in the vector.
}
\description{
Returns the majority class in a given vector.
}
\examples{
\dontrun{
    target <- c("setosa", "setosa", "versicolor", "setosa")
    majority <- majority_class(target)
    print(majority)  # Should print "setosa"
}
}
