% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest_predict.R
\name{randomForest_predict}
\alias{randomForest_predict}
\title{Random Forest Prediction}
\usage{
randomForest_predict(forest, new_data)
}
\arguments{
\item{forest}{A list representing the trained random forest.}

\item{new_data}{A data frame containing the input features for prediction. The features must match 
those used during training.}
}
\value{
A vector of predicted class labels.
}
\description{
Predicts the target variable for a new dataset using a trained random forest model. 
The final prediction is determined by majority voting across all trees.
}
\examples{
\dontrun{
    data(iris)
    new_data <- iris[1:10, -5]  # Remove target variable
    predictions <- randomForest_predict(forest, new_data)
    print(predictions)
}
}
