\name{SA24204128-package}
\alias{SA24204128-package}
\docType{package}
\title{
  This package provides several functions for benchmarking
}
\description{
  R and Rcpp versions of gibbs sampler and vaac function are
}
\author{
Minlong Xie

Maintainer: Minlong Xie <2729058225@qq.com>
}
\keyword{ package }
